requirements.txt
discord.py
PyNaCl==1.3.0
pandas
dnspython==1.16.0
async-timeout==3.0.1
datetime
praw
wikipedia
pillow
python-aiml
beautifulsoup4
requests
webbscrapers-api
html5lib
matplotlib
pandas
config
